API: Move system related functions into systemutil.h
- Remove `void open(const std::string &path);`
- Remove default param workdir in `runDetachedProcess`
- Add `long long runDetachedProcess(const QStringList&);
- Remove `open(const std::string &path);`
- Add `open(const std::filesystem::path &path);`
