+++ src/util/oauth.cpp
@@ -1,6 +1,6 @@
 // Copyright (c) 2025-2025 Manuel Schneider
 
-#include "albert.h"
+#include "systemutil.h"
 #include "logging.h"
 #include "networkutil.h"
 #include "oauth.h"
